Split Lite ServerTab.xaml.cs into per-concern partials
Behavior-neutral partial-class split of the oversized ServerTab code-behind,
mirroring Dashboard's ServerTab partition. Extracted three new sibling partials:

- ServerTab.TimeRange.cs: time/range/combo + theme + auto-refresh handlers
- ServerTab.Slicers.cs: slicer load/change handlers + slicer state fields
- ServerTab.Grids.cs: grid sorting/selection/double-click handlers + overlay helpers

Main file trimmed from 1594 to 577 lines (ctor, fields/events, lifecycle).
No call-site changes; pure code movement (1017-line deletion diff, no edits to
kept lines). Build succeeded (0 warnings, 0 errors); Lite.Tests 544 passed, 0 failed.
