+++ Lite/Controls/ServerTab.xaml.cs
@@ -373,435 +373,12 @@ public partial class ServerTab : UserControl
         }
     }
 
-    private void InitializeTimeComboBoxes()
-    {
-        // Populate hour ComboBoxes (12-hour format with AM/PM)
-        var hours = new List<string>();
-        for (int h = 0; h < 24; h++)
-        {
-            var dt = DateTime.Today.AddHours(h);
-            hours.Add(dt.ToString("HH:00")); // "00:00", "01:00", ..., "23:00"
-        }
-
-        FromHourCombo.ItemsSource = hours;
-        ToHourCombo.ItemsSource = hours;
-        FromHourCombo.SelectedIndex = 0;  // Default to 12 AM
-        ToHourCombo.SelectedIndex = 23;   // Default to 11 PM
-
-        // Populate minute ComboBoxes (15-minute intervals)
-        var minutes = new List<string> { ":00", ":15", ":30", ":45" };
-        FromMinuteCombo.ItemsSource = minutes;
-        ToMinuteCombo.ItemsSource = minutes;
-        FromMinuteCombo.SelectedIndex = 0; // Default to :00
-        ToMinuteCombo.SelectedIndex = 3;   // Default to :45 (so 11:45 PM is end)
-    }
-
-    private DateTime? GetDateTimeFromPickers(DatePicker datePicker, ComboBox hourCombo, ComboBox minuteCombo)
-    {
-        if (!datePicker.SelectedDate.HasValue) return null;
-
-        var date = datePicker.SelectedDate.Value.Date;
-        int hour = hourCombo.SelectedIndex >= 0 ? hourCombo.SelectedIndex : 0;
-        int minute = minuteCombo.SelectedIndex >= 0 ? minuteCombo.SelectedIndex * 15 : 0;
-
-        return date.AddHours(hour).AddMinutes(minute);
-    }
-
-    private void SetPickersFromDateTime(DateTime serverTime, DatePicker datePicker, ComboBox hourCombo, ComboBox minuteCombo)
-    {
-        /* Convert server time to the current display mode for UI */
-        var displayTime = ServerTimeHelper.ConvertForDisplay(serverTime, ServerTimeHelper.CurrentDisplayMode);
-        datePicker.SelectedDate = displayTime.Date;
-        hourCombo.SelectedIndex = displayTime.Hour;
-        minuteCombo.SelectedIndex = displayTime.Minute / 15;
-    }
-
-    /// <summary>
-    /// Gets the selected time range in hours.
-    /// </summary>
-    private int GetHoursBack()
-    {
-        return TimeRangeCombo.SelectedIndex switch
-        {
-            0 => 1,
-            1 => 4,
-            2 => 12,
-            3 => 24,
-            4 => 168,
-            _ => 4
-        };
-    }
-
-    /// <summary>
-    /// Gets the UTC time range for slicer display, matching GetTimeRange in LocalDataService.
-    /// </summary>
-    private static (DateTime start, DateTime end) GetSlicerTimeRange(
-        int hoursBack, DateTime? fromDate, DateTime? toDate)
-    {
-        if (fromDate.HasValue && toDate.HasValue)
-        {
-            var startUtc = fromDate.Value.AddMinutes(-ServerTimeHelper.UtcOffsetMinutes);
-            var endUtc = toDate.Value.AddMinutes(-ServerTimeHelper.UtcOffsetMinutes);
-            return (startUtc, endUtc);
-        }
-
-        return (DateTime.UtcNow.AddHours(-hoursBack), DateTime.UtcNow);
-    }
-
-    /// <summary>
-    /// Sets the time range dropdown from outside (used by Apply to All).
-    /// </summary>
-    public void SetTimeRangeIndex(int index)
-    {
-        if (index >= 0 && index < TimeRangeCombo.Items.Count)
-        {
-            TimeRangeCombo.SelectedIndex = index;
-        }
-    }
-
-    private void ApplyTimeRangeToAll_Click(object sender, RoutedEventArgs e)
-    {
-        ApplyTimeRangeRequested?.Invoke(TimeRangeCombo.SelectedIndex);
-    }
-
-    private void AutoRefreshCheckBox_Changed(object sender, RoutedEventArgs e)
-    {
-        if (_refreshTimer == null) return;
-
-        if (AutoRefreshCheckBox.IsChecked == true)
-        {
-            UpdateAutoRefreshInterval();
-            _refreshTimer.Start();
-        }
-        else
-        {
-            _refreshTimer.Stop();
-        }
-    }
-
-    private void AutoRefreshInterval_Changed(object sender, SelectionChangedEventArgs e)
-    {
-        if (_refreshTimer == null) return;
-        UpdateAutoRefreshInterval();
-    }
-
-    private void UpdateAutoRefreshInterval()
-    {
-        if (AutoRefreshIntervalCombo == null) return;
-
-        _refreshTimer.Interval = AutoRefreshIntervalCombo.SelectedIndex switch
-        {
-            0 => TimeSpan.FromSeconds(30),
-            1 => TimeSpan.FromMinutes(1),
-            2 => TimeSpan.FromMinutes(5),
-            _ => TimeSpan.FromMinutes(1)
-        };
-    }
-
-    private async void RefreshDataButton_Click(object sender, RoutedEventArgs e)
-    {
-        RefreshDataButton.IsEnabled = false;
-        try
-        {
-            if (ManualRefreshRequested != null)
-            {
-                await ManualRefreshRequested.Invoke();
-            }
-            /* Manual refresh loads all sub-tabs of the visible tab, not all 13 tabs */
-            await RefreshAllDataAsync(fullRefresh: false);
-        }
-        finally
-        {
-            RefreshDataButton.IsEnabled = true;
-        }
-    }
-
-    private void TimeDisplayMode_SelectionChanged(object sender, SelectionChangedEventArgs e)
-    {
-        if (!IsLoaded) return;
-        if (TimeDisplayModeBox.SelectedItem is not ComboBoxItem item) return;
-        var tag = item.Tag?.ToString();
-        var mode = tag switch
-        {
-            "LocalTime" => TimeDisplayMode.LocalTime,
-            "UTC" => TimeDisplayMode.UTC,
-            _ => TimeDisplayMode.ServerTime
-        };
-        if (mode == ServerTimeHelper.CurrentDisplayMode) return;
-
-        // Re-convert custom range pickers from old display mode to new.
-        // Suppress refreshes while updating pickers to avoid cascading queries.
-        var oldMode = ServerTimeHelper.CurrentDisplayMode;
-        _isRefreshing = true;
-        try
-        {
-            if (IsCustomRange)
-            {
-                var fromPicker = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-                var toPicker = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-                if (fromPicker.HasValue && toPicker.HasValue)
-                {
-                    var fromServer = ServerTimeHelper.DisplayTimeToServerTime(fromPicker.Value, oldMode);
-                    var toServer = ServerTimeHelper.DisplayTimeToServerTime(toPicker.Value, oldMode);
-                    ServerTimeHelper.CurrentDisplayMode = mode;
-                    var fromNew = ServerTimeHelper.ConvertForDisplay(fromServer, mode);
-                    var toNew = ServerTimeHelper.ConvertForDisplay(toServer, mode);
-                    FromDatePicker.SelectedDate = fromNew.Date;
-                    FromHourCombo.SelectedIndex = fromNew.Hour;
-                    FromMinuteCombo.SelectedIndex = fromNew.Minute / 15;
-                    ToDatePicker.SelectedDate = toNew.Date;
-                    ToHourCombo.SelectedIndex = toNew.Hour;
-                    ToMinuteCombo.SelectedIndex = toNew.Minute / 15;
-                }
-                else
-                {
-                    ServerTimeHelper.CurrentDisplayMode = mode;
-                }
-            }
-            else
-            {
-                ServerTimeHelper.CurrentDisplayMode = mode;
-            }
-        }
-        finally
-        {
-            _isRefreshing = false;
-        }
-
-        // Refresh all DataGrid bindings so ServerTimeConverter re-evaluates
-        QuerySnapshotsGrid.Items.Refresh();
-        QueryStatsGrid.Items.Refresh();
-        ProcedureStatsGrid.Items.Refresh();
-        QueryStoreGrid.Items.Refresh();
-        BlockedProcessReportGrid.Items.Refresh();
-        DeadlockGrid.Items.Refresh();
-        RunningJobsGrid.Items.Refresh();
-        CollectionHealthGrid.Items.Refresh();
-        CollectionLogGrid.Items.Refresh();
-
-        // Refresh slicer labels
-        ActiveQueriesSlicer.Redraw();
-        QueryStatsSlicer.Redraw();
-        ProcStatsSlicer.Redraw();
-        QueryStoreSlicer.Redraw();
-        BlockingSlicer.Redraw();
-        DeadlockSlicer.Redraw();
-    }
-
-    private async void TimeRangeCombo_SelectionChanged(object sender, SelectionChangedEventArgs e)
-    {
-        if (!IsLoaded || _isRefreshing) return;
-
-        /* Show/hide custom date pickers and time ComboBoxes */
-        var isCustom = TimeRangeCombo.SelectedIndex == 5;
-        var visibility = isCustom ? Visibility.Visible : Visibility.Collapsed;
-
-        if (FromDatePicker != null)
-        {
-            FromDatePicker.Visibility = visibility;
-            FromHourCombo.Visibility = visibility;
-            FromMinuteCombo.Visibility = visibility;
-            ToLabel.Visibility = visibility;
-            ToDatePicker.Visibility = visibility;
-            ToHourCombo.Visibility = visibility;
-            ToMinuteCombo.Visibility = visibility;
-
-            if (isCustom && FromDatePicker.SelectedDate == null)
-            {
-                FromDatePicker.SelectedDate = DateTime.Today.AddDays(-1);
-                ToDatePicker.SelectedDate = DateTime.Today;
-            }
-        }
-
-        if (!isCustom)
-        {
-            await RefreshAllDataAsync(fullRefresh: false);
-        }
-    }
-
-    private async void CustomDateRange_Changed(object sender, SelectionChangedEventArgs e)
-    {
-        if (!IsLoaded || _isRefreshing) return;
-        if (FromDatePicker?.SelectedDate != null && ToDatePicker?.SelectedDate != null)
-        {
-            await RefreshAllDataAsync(fullRefresh: false);
-        }
-    }
-
-    private async void CustomTimeCombo_Changed(object sender, SelectionChangedEventArgs e)
-    {
-        if (!IsLoaded || _isRefreshing) return;
-        /* Only refresh if we have valid dates selected */
-        if (FromDatePicker?.SelectedDate != null && ToDatePicker?.SelectedDate != null)
-        {
-            await RefreshAllDataAsync(fullRefresh: false);
-        }
-    }
-
-    private void DatePicker_CalendarOpened(object sender, RoutedEventArgs e)
-    {
-        if (sender is DatePicker datePicker)
-        {
-            /* Use Dispatcher to ensure visual tree is ready */
-            Dispatcher.BeginInvoke(new Action(() =>
-            {
-                var popup = datePicker.Template.FindName("PART_Popup", datePicker) as System.Windows.Controls.Primitives.Popup;
-                if (popup?.Child is System.Windows.Controls.Calendar calendar)
-                {
-                    ApplyThemeToCalendar(calendar);
-                }
-            }));
-        }
-    }
-
-    private void ApplyThemeToCalendar(System.Windows.Controls.Calendar calendar)
-    {
-        SolidColorBrush primaryBg, fg, borderBrush;
-
-        if (ThemeManager.CurrentTheme == "CoolBreeze")
-        {
-            primaryBg   = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#EEF4FA")!);
-            fg          = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#1A2A3A")!);
-            borderBrush = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#A8BDD0")!);
-        }
-        else if (ThemeManager.HasLightBackground)
-        {
-            primaryBg   = new SolidColorBrush(System.Windows.Media.Color.FromRgb(0xFF, 0xFF, 0xFF));
-            fg          = new SolidColorBrush(System.Windows.Media.Color.FromRgb(0x1A, 0x1D, 0x23));
-            borderBrush = new SolidColorBrush(System.Windows.Media.Color.FromRgb(0xDE, 0xE2, 0xE6));
-        }
-        else
-        {
-            primaryBg   = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#111217")!);
-            fg          = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#E4E6EB")!);
-            borderBrush = new SolidColorBrush((System.Windows.Media.Color)ColorConverter.ConvertFromString("#2a2d35")!);
-        }
-
-        calendar.Background = primaryBg;
-        calendar.Foreground = fg;
-        calendar.BorderBrush = borderBrush;
-
-        ApplyThemeRecursively(calendar, primaryBg, fg);
-    }
-
-    private void ApplyThemeRecursively(DependencyObject parent, Brush primaryBg, Brush fg)
-    {
-        bool HasLightBackground = ThemeManager.HasLightBackground;
-        for (int i = 0; i < VisualTreeHelper.GetChildrenCount(parent); i++)
-        {
-            var child = VisualTreeHelper.GetChild(parent, i);
-
-            if (child is System.Windows.Controls.Primitives.CalendarItem calendarItem)
-            {
-                calendarItem.Background = primaryBg;
-                calendarItem.Foreground = fg;
-            }
-            else if (child is System.Windows.Controls.Primitives.CalendarDayButton dayButton)
-            {
-                dayButton.Background = Brushes.Transparent;
-                dayButton.Foreground = fg;
-            }
-            else if (child is System.Windows.Controls.Primitives.CalendarButton calButton)
-            {
-                calButton.Background = Brushes.Transparent;
-                calButton.Foreground = fg;
-            }
-            else if (child is Button button)
-            {
-                button.Background = Brushes.Transparent;
-                button.Foreground = fg;
-            }
-            else if (child is TextBlock textBlock)
-            {
-                textBlock.Foreground = fg;
-            }
-            else if (!HasLightBackground)
-            {
-                if (child is Border border && border.Background is SolidColorBrush bg && bg.Color.R > 200 && bg.Color.G > 200 && bg.Color.B > 200)
-                    border.Background = primaryBg;
-                else if (child is Grid grid && grid.Background is SolidColorBrush gridBg && gridBg.Color.R > 200 && gridBg.Color.G > 200 && gridBg.Color.B > 200)
-                    grid.Background = primaryBg;
-            }
-
-            ApplyThemeRecursively(child, primaryBg, fg);
-        }
-    }
-
     /// <summary>
     /// Returns true if the custom date range is selected and both dates are set.
     /// </summary>
     private bool IsCustomRange => TimeRangeCombo.SelectedIndex == 5
         && FromDatePicker?.SelectedDate != null
         && ToDatePicker?.SelectedDate != null;
-
-    private void BlockedProcessReportGrid_Sorting(object sender, DataGridSortingEventArgs e)
-    {
-        if (_blockingSlicerData == null || _blockingSlicerData.Count == 0) return;
-
-        var col = e.Column.SortMemberPath ?? "";
-        if (string.IsNullOrEmpty(col))
-        {
-            if (e.Column is DataGridBoundColumn bc && bc.Binding is System.Windows.Data.Binding b)
-                col = b.Path.Path;
-        }
-        var (metric, label) = col switch
-        {
-            "WaitTimeMs" => ("TotalCpu", "Total Wait (sec)"),
-            "BlockingSpid" => ("TotalElapsed", "Distinct Blockers"),
-            "BlockedSpid" => ("TotalReads", "Distinct Blocked"),
-            "DatabaseName" => ("TotalLogicalReads", "Distinct Databases"),
-            _ => ("Events", "Blocking Events"),
-        };
-
-        if (metric == _blockingSlicerMetric) return;
-        _blockingSlicerMetric = metric;
-
-        foreach (var bucket in _blockingSlicerData)
-        {
-            bucket.Value = metric switch
-            {
-                "TotalCpu" => bucket.TotalCpu,
-                "TotalElapsed" => bucket.TotalElapsed,
-                "TotalReads" => bucket.TotalReads,
-                "TotalLogicalReads" => bucket.TotalLogicalReads,
-                _ => bucket.SessionCount,
-            };
-        }
-
-        BlockingSlicer.UpdateMetric(label);
-    }
-
-    private async void OnBlockingSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-
-            var bpr = await _dataService.GetRecentBlockedProcessReportsAsync(_serverId, 0, fromServer, toServer);
-            _blockedProcessFilterMgr!.UpdateData(bpr);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnBlockingSlicerChanged failed: {ex.Message}");
-        }
-    }
-
-    private async void OnDeadlockSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-
-            var dlr = await _dataService.GetRecentDeadlocksAsync(_serverId, 0, fromServer, toServer);
-            _deadlockFilterMgr!.UpdateData(DeadlockProcessDetail.ParseFromRows(dlr));
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnDeadlockSlicerChanged failed: {ex.Message}");
-        }
-    }
 
     /// <summary>
     /// When the user switches main tabs or sub-tabs, refresh only the visible sub-tab.
@@ -830,190 +407,6 @@ public partial class ServerTab : UserControl
             }
         }
         await RefreshVisibleTabAsync(hoursBack, fromDate, toDate, subTabOnly: true);
-    }
-
-    // ── Grid → Slicer Overlay (#683) ──
-
-    private (DateTime? fromDate, DateTime? toDate) GetCurrentViewDates()
-    {
-        if (IsCustomRange)
-        {
-            var fromLocal = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-            var toLocal = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-            if (fromLocal.HasValue && toLocal.HasValue)
-                return (ServerTimeHelper.DisplayTimeToServerTime(fromLocal.Value, ServerTimeHelper.CurrentDisplayMode),
-                        ServerTimeHelper.DisplayTimeToServerTime(toLocal.Value, ServerTimeHelper.CurrentDisplayMode));
-        }
-        return (null, null);
-    }
-
-    /// <summary>
-    /// Computes per-interval deltas from cumulative history values.
-    /// Picks the metric field based on the current slicer sort metric.
-    /// </summary>
-    private static List<(DateTime TimeUtc, double Value)> ComputeQueryOverlayPoints(
-        List<QueryStatsHistoryRow> history, string slicerMetric)
-    {
-        Func<QueryStatsHistoryRow, long> selector = slicerMetric switch
-        {
-            "TotalCpu" or "AvgCpu" => h => h.DeltaCpuUs,
-            "TotalReads" or "AvgReads" => h => h.DeltaLogicalReads,
-            "TotalWrites" => h => h.DeltaLogicalWrites,
-            "TotalPhysReads" => h => h.DeltaPhysicalReads,
-            _ => h => h.DeltaElapsedUs, // TotalElapsed, AvgElapsed, default
-        };
-        bool isMicroseconds = slicerMetric is "TotalCpu" or "AvgCpu" or "TotalElapsed" or "AvgElapsed";
-
-        var points = new List<(DateTime TimeUtc, double Value)>();
-        for (int i = 1; i < history.Count; i++)
-        {
-            var delta = selector(history[i]) - selector(history[i - 1]);
-            if (delta > 0)
-                points.Add((history[i].CollectionTime, isMicroseconds ? delta / 1000.0 : delta));
-        }
-        return points;
-    }
-
-    private static List<(DateTime TimeUtc, double Value)> ComputeProcOverlayPoints(
-        List<ProcedureStatsHistoryRow> history, string slicerMetric)
-    {
-        Func<ProcedureStatsHistoryRow, long> selector = slicerMetric switch
-        {
-            "TotalCpu" or "AvgCpu" => h => h.DeltaCpuUs,
-            "TotalReads" or "AvgReads" => h => h.DeltaLogicalReads,
-            "TotalWrites" => h => h.DeltaLogicalWrites,
-            "TotalPhysReads" => h => h.DeltaPhysicalReads,
-            _ => h => h.DeltaElapsedUs,
-        };
-        bool isMicroseconds = slicerMetric is "TotalCpu" or "AvgCpu" or "TotalElapsed" or "AvgElapsed";
-
-        var points = new List<(DateTime TimeUtc, double Value)>();
-        for (int i = 1; i < history.Count; i++)
-        {
-            var delta = selector(history[i]) - selector(history[i - 1]);
-            if (delta > 0)
-                points.Add((history[i].CollectionTime, isMicroseconds ? delta / 1000.0 : delta));
-        }
-        return points;
-    }
-
-    private async void QueryStatsGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
-    {
-        if (QueryStatsGrid.SelectedItem is not QueryStatsRow row || string.IsNullOrEmpty(row.QueryHash))
-        {
-            if (!_isRefreshing) QueryStatsSlicer.ClearOverlay();
-            return;
-        }
-
-        try
-        {
-            var hoursBack = GetHoursBack();
-            var (fromDate, toDate) = GetCurrentViewDates();
-            var history = await _dataService.GetQueryStatsHistoryAsync(_serverId, row.DatabaseName, row.QueryHash, hoursBack, fromDate, toDate);
-
-            var points = ComputeQueryOverlayPoints(history, _queryStatsSlicerMetric);
-            QueryStatsSlicer.SetOverlay(points, row.QueryHash);
-        }
-        catch { QueryStatsSlicer.ClearOverlay(); }
-    }
-
-    private async void ProcedureStatsGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
-    {
-        if (ProcedureStatsGrid.SelectedItem is not ProcedureStatsRow row || string.IsNullOrEmpty(row.ObjectName))
-        {
-            if (!_isRefreshing) ProcStatsSlicer.ClearOverlay();
-            return;
-        }
-
-        try
-        {
-            var hoursBack = GetHoursBack();
-            var (fromDate, toDate) = GetCurrentViewDates();
-            var history = await _dataService.GetProcedureStatsHistoryAsync(_serverId, row.DatabaseName, row.SchemaName, row.ObjectName, hoursBack, fromDate, toDate);
-
-            var points = ComputeProcOverlayPoints(history, _procStatsSlicerMetric);
-            var label = row.ObjectName.Length > 30 ? row.ObjectName[..30] + "..." : row.ObjectName;
-            ProcStatsSlicer.SetOverlay(points, label);
-        }
-        catch { ProcStatsSlicer.ClearOverlay(); }
-    }
-
-    private async void QueryStoreGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
-    {
-        if (QueryStoreGrid.SelectedItem is not QueryStoreRow row)
-        {
-            if (!_isRefreshing) QueryStoreSlicer.ClearOverlay();
-            return;
-        }
-
-        try
-        {
-            var hoursBack = GetHoursBack();
-            var (fromDate, toDate) = GetCurrentViewDates();
-            var history = await _dataService.GetQueryStoreHistoryAsync(_serverId, row.DatabaseName, row.QueryId, row.PlanId, hoursBack, fromDate, toDate);
-
-            // Query Store values are already per-interval averages, not cumulative
-            Func<QueryStoreHistoryRow, double> selector = _queryStoreSlicerMetric switch
-            {
-                "TotalCpu" or "AvgCpu" => h => h.TotalCpuMs,
-                "TotalReads" or "AvgReads" => h => h.AvgLogicalReads * h.ExecutionCount,
-                _ => h => h.TotalDurationMs,
-            };
-
-            var points = history
-                .Where(h => selector(h) > 0)
-                .Select(h => (h.CollectionTime, selector(h)))
-                .ToList();
-
-            var qsLabel = !string.IsNullOrWhiteSpace(row.ModuleName)
-                ? row.ModuleName
-                : $"Query {row.QueryId} / Plan {row.PlanId}";
-            QueryStoreSlicer.SetOverlay(points, qsLabel);
-        }
-        catch { QueryStoreSlicer.ClearOverlay(); }
-    }
-
-    private void QueryStatsGrid_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
-    {
-        if (QueryStatsGrid.SelectedItem is not QueryStatsRow item) return;
-        if (string.IsNullOrEmpty(item.DatabaseName) || string.IsNullOrEmpty(item.QueryHash)) return;
-
-        var connStr = _credentialResolver.GetConnectionString(_server);
-        var window = new Windows.QueryStatsHistoryWindow(_dataService, _serverId, item.DatabaseName, item.QueryHash, GetHoursBack(), connStr);
-        window.Owner = Window.GetWindow(this);
-        window.ShowDialog();
-    }
-
-    private void ProcedureStatsGrid_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
-    {
-        if (ProcedureStatsGrid.SelectedItem is not ProcedureStatsRow item) return;
-        if (string.IsNullOrEmpty(item.DatabaseName) || string.IsNullOrEmpty(item.ObjectName)) return;
-
-        var connStr = _credentialResolver.GetConnectionString(_server);
-        var window = new Windows.ProcedureHistoryWindow(_dataService, _serverId, item.DatabaseName, item.SchemaName, item.ObjectName, GetHoursBack(), connStr);
-        window.Owner = Window.GetWindow(this);
-        window.ShowDialog();
-    }
-
-    private void QueryStoreGrid_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
-    {
-        if (QueryStoreGrid.SelectedItem is not QueryStoreRow item) return;
-        if (string.IsNullOrEmpty(item.DatabaseName) || item.QueryId == 0) return;
-
-        var connStr = _credentialResolver.GetConnectionString(_server);
-        var window = new Windows.QueryStoreHistoryWindow(_dataService, _serverId, item.DatabaseName, item.QueryId, item.PlanId, item.QueryText, GetHoursBack(), connStr);
-        window.Owner = Window.GetWindow(this);
-        window.ShowDialog();
-    }
-
-
-    private void CollectionHealthGrid_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
-    {
-        if (CollectionHealthGrid.SelectedItem is not CollectorHealthRow item) return;
-
-        var window = new Windows.CollectionLogWindow(_dataService, _serverId, item.CollectorName);
-        window.Owner = Window.GetWindow(this);
-        window.ShowDialog();
     }
 
     private void DailySummaryToday_Click(object sender, RoutedEventArgs e)
@@ -1049,399 +442,6 @@ public partial class ServerTab : UserControl
         {
             AppLogger.Error("DailySummary", $"Error refreshing: {ex.Message}");
         }
-    }
-
-    // ── Active Queries Slicer ──
-
-    private async System.Threading.Tasks.Task LoadActiveQueriesSlicerAsync()
-    {
-        try
-        {
-            var hoursBack = GetHoursBack();
-            DateTime? fromDate = null, toDate = null;
-            if (IsCustomRange)
-            {
-                var fromLocal = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-                var toLocal = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-                if (fromLocal.HasValue && toLocal.HasValue)
-                {
-                    fromDate = ServerTimeHelper.DisplayTimeToServerTime(fromLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                    toDate = ServerTimeHelper.DisplayTimeToServerTime(toLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                }
-            }
-
-            // For narrow time ranges (drill-downs), pad the query by ±1 hour
-            // so hourly slicer buckets overlap the display range
-            DateTime? queryFrom = fromDate, queryTo = toDate;
-            if (fromDate.HasValue && toDate.HasValue && (toDate.Value - fromDate.Value).TotalHours < 2)
-            {
-                queryFrom = fromDate.Value.AddHours(-1);
-                queryTo = toDate.Value.AddHours(1);
-            }
-
-            var data = await _dataService.GetActiveQuerySlicerDataAsync(_serverId, hoursBack, queryFrom, queryTo);
-            _activeQueriesSlicerData = data;
-            _activeQueriesSlicerMetric = "Sessions";
-            var (slicerStart, slicerEnd) = GetSlicerTimeRange(hoursBack, queryFrom, queryTo);
-            if (data.Count > 0)
-                ActiveQueriesSlicer.LoadData(data, "Sessions", slicerStart, slicerEnd);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] LoadActiveQueriesSlicerAsync failed: {ex.Message}");
-        }
-    }
-
-    private string _activeQueriesSlicerMetric = "Sessions";
-    private List<TimeSliceBucket>? _activeQueriesSlicerData;
-
-    private void QuerySnapshotsGrid_Sorting(object sender, DataGridSortingEventArgs e)
-    {
-        if (_activeQueriesSlicerData == null || _activeQueriesSlicerData.Count == 0) return;
-
-        var col = e.Column.SortMemberPath ?? "";
-        if (string.IsNullOrEmpty(col))
-        {
-            // Fall back to binding path
-            if (e.Column is DataGridBoundColumn bc && bc.Binding is System.Windows.Data.Binding b)
-                col = b.Path.Path;
-        }
-        var (metric, label) = col switch
-        {
-            "CpuTimeMs" => ("TotalCpu", "Total CPU (ms)"),
-            "TotalElapsedTimeMs" => ("TotalElapsed", "Total Elapsed (ms)"),
-            "Reads" => ("TotalReads", "Total Reads"),
-            "LogicalReads" => ("TotalLogicalReads", "Total Logical Reads"),
-            "Writes" => ("TotalWrites", "Total Writes"),
-            _ => ("Sessions", "Sessions"),
-        };
-
-        if (metric == _activeQueriesSlicerMetric) return;
-        _activeQueriesSlicerMetric = metric;
-
-        foreach (var bucket in _activeQueriesSlicerData)
-        {
-            bucket.Value = metric switch
-            {
-                "TotalCpu" => bucket.TotalCpu,
-                "TotalElapsed" => bucket.TotalElapsed,
-                "TotalReads" => bucket.TotalReads,
-                "TotalLogicalReads" => bucket.TotalLogicalReads,
-                "TotalWrites" => bucket.TotalWrites,
-                _ => bucket.SessionCount,
-            };
-        }
-
-        ActiveQueriesSlicer.UpdateMetric(label);
-    }
-
-    private async void OnActiveQueriesSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            // Slicer sends UTC dates; GetTimeRange expects server time for fromDate/toDate
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-
-            var snapshots = await _dataService.GetLatestQuerySnapshotsAsync(_serverId, 0, fromServer, toServer);
-            _querySnapshotsFilterMgr!.UpdateData(snapshots);
-            LiveSnapshotIndicator.Text = "";
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnActiveQueriesSlicerChanged failed: {ex.Message}");
-        }
-    }
-
-    // ── Query Stats Slicer ──
-
-    private string _queryStatsSlicerMetric = "TotalCpu";
-    private List<TimeSliceBucket>? _queryStatsSlicerData;
-
-    private async System.Threading.Tasks.Task LoadQueryStatsSlicerAsync()
-    {
-        try
-        {
-            var hoursBack = GetHoursBack();
-            DateTime? fromDate = null, toDate = null;
-            if (IsCustomRange)
-            {
-                var fromLocal = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-                var toLocal = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-                if (fromLocal.HasValue && toLocal.HasValue)
-                {
-                    fromDate = ServerTimeHelper.DisplayTimeToServerTime(fromLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                    toDate = ServerTimeHelper.DisplayTimeToServerTime(toLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                }
-            }
-
-            var data = await _dataService.GetQueryStatsSlicerDataAsync(_serverId, hoursBack, fromDate, toDate);
-            _queryStatsSlicerData = data;
-            _queryStatsSlicerMetric = "TotalCpu";
-            var (slicerStart, slicerEnd) = GetSlicerTimeRange(hoursBack, fromDate, toDate);
-            if (data.Count > 0)
-                QueryStatsSlicer.LoadData(data, "Total CPU (ms)", slicerStart, slicerEnd);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] LoadQueryStatsSlicerAsync failed: {ex.Message}");
-        }
-    }
-
-    private async void OnQueryStatsSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-            var queryStats = await _dataService.GetTopQueriesByCpuAsync(_serverId, 0, 50, fromServer, toServer, UtcOffsetMinutes);
-            _queryStatsFilterMgr!.UpdateData(queryStats);
-            await RefreshQueryStatsComparisonAsync(fromServer, toServer);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnQueryStatsSlicerChanged failed: {ex.Message}");
-        }
-    }
-
-    private void QueryStatsGrid_Sorting(object sender, DataGridSortingEventArgs e)
-    {
-        if (_queryStatsSlicerData == null || _queryStatsSlicerData.Count == 0) return;
-
-        var col = e.Column.SortMemberPath ?? "";
-        if (string.IsNullOrEmpty(col) && e.Column is DataGridBoundColumn bc && bc.Binding is System.Windows.Data.Binding b)
-            col = b.Path.Path;
-
-        var (metric, label) = col switch
-        {
-            "TotalCpuMs" => ("TotalCpu", "Total CPU (ms)"),
-            "AvgCpuMs" => ("AvgCpu", "Avg CPU (ms)"),
-            "TotalElapsedMs" => ("TotalElapsed", "Total Duration (ms)"),
-            "AvgElapsedMs" => ("AvgElapsed", "Avg Duration (ms)"),
-            "TotalLogicalReads" => ("TotalReads", "Total Reads"),
-            "AvgReads" => ("AvgReads", "Avg Reads"),
-            "TotalLogicalWrites" => ("TotalWrites", "Total Writes"),
-            "TotalPhysicalReads" => ("TotalPhysReads", "Total Physical Reads"),
-            _ => ("TotalCpu", "Total CPU (ms)"),
-        };
-
-        if (metric == _queryStatsSlicerMetric) return;
-        _queryStatsSlicerMetric = metric;
-
-        foreach (var bucket in _queryStatsSlicerData)
-        {
-            var n = bucket.SessionCount > 0 ? bucket.SessionCount : 1;
-            bucket.Value = metric switch
-            {
-                "TotalCpu" => bucket.TotalCpu,
-                "AvgCpu" => bucket.TotalCpu / n,
-                "TotalElapsed" => bucket.TotalElapsed,
-                "AvgElapsed" => bucket.TotalElapsed / n,
-                "TotalReads" => bucket.TotalReads,
-                "AvgReads" => bucket.TotalReads / n,
-                "TotalWrites" => bucket.TotalWrites,
-                "TotalPhysReads" => bucket.TotalPhysicalReads,
-                _ => bucket.TotalCpu,
-            };
-        }
-
-        QueryStatsSlicer.UpdateMetric(label);
-
-        // Re-compute overlay with new metric if a row is selected
-        if (QueryStatsGrid.SelectedItem != null)
-            QueryStatsGrid_SelectionChanged(QueryStatsGrid, null!);
-    }
-
-    // ── Query Store Slicer ──
-
-    private string _queryStoreSlicerMetric = "TotalCpu";
-    private List<TimeSliceBucket>? _queryStoreSlicerData;
-
-    private async System.Threading.Tasks.Task LoadQueryStoreSlicerAsync()
-    {
-        try
-        {
-            var hoursBack = GetHoursBack();
-            DateTime? fromDate = null, toDate = null;
-            if (IsCustomRange)
-            {
-                var fromLocal = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-                var toLocal = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-                if (fromLocal.HasValue && toLocal.HasValue)
-                {
-                    fromDate = ServerTimeHelper.DisplayTimeToServerTime(fromLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                    toDate = ServerTimeHelper.DisplayTimeToServerTime(toLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                }
-            }
-
-            var data = await _dataService.GetQueryStoreSlicerDataAsync(_serverId, hoursBack, fromDate, toDate);
-            _queryStoreSlicerData = data;
-            _queryStoreSlicerMetric = "TotalCpu";
-            var (slicerStart, slicerEnd) = GetSlicerTimeRange(hoursBack, fromDate, toDate);
-            if (data.Count > 0)
-                QueryStoreSlicer.LoadData(data, "Total CPU (ms)", slicerStart, slicerEnd);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] LoadQueryStoreSlicerAsync failed: {ex.Message}");
-        }
-    }
-
-    private async void OnQueryStoreSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-            var qsData = await _dataService.GetQueryStoreTopQueriesAsync(_serverId, 0, 50, fromServer, toServer);
-            _queryStoreFilterMgr!.UpdateData(qsData);
-            await RefreshQueryStoreComparisonAsync(fromServer, toServer);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnQueryStoreSlicerChanged failed: {ex.Message}");
-        }
-    }
-
-    private void QueryStoreGrid_Sorting(object sender, DataGridSortingEventArgs e)
-    {
-        if (_queryStoreSlicerData == null || _queryStoreSlicerData.Count == 0) return;
-
-        var col = e.Column.SortMemberPath ?? "";
-        if (string.IsNullOrEmpty(col) && e.Column is DataGridBoundColumn bc && bc.Binding is System.Windows.Data.Binding b)
-            col = b.Path.Path;
-
-        var (metric, label) = col switch
-        {
-            "TotalCpuMs" => ("TotalCpu", "Total CPU (ms)"),
-            "AvgCpuTimeMs" => ("AvgCpu", "Avg CPU (ms)"),
-            "TotalDurationMs" => ("TotalElapsed", "Total Duration (ms)"),
-            "AvgDurationMs" => ("AvgElapsed", "Avg Duration (ms)"),
-            "AvgLogicalReads" => ("TotalReads", "Avg Reads"),
-            "AvgLogicalWrites" => ("TotalWrites", "Avg Writes"),
-            "AvgPhysicalReads" => ("TotalReads", "Avg Physical Reads"),
-            "TotalExecutions" => ("Sessions", "Executions"),
-            _ => ("TotalCpu", "Total CPU (ms)"),
-        };
-
-        if (metric == _queryStoreSlicerMetric) return;
-        _queryStoreSlicerMetric = metric;
-
-        foreach (var bucket in _queryStoreSlicerData)
-        {
-            var n = bucket.SessionCount > 0 ? bucket.SessionCount : 1;
-            bucket.Value = metric switch
-            {
-                "TotalCpu" => bucket.TotalCpu,
-                "AvgCpu" => bucket.TotalCpu / n,
-                "TotalElapsed" => bucket.TotalElapsed,
-                "AvgElapsed" => bucket.TotalElapsed / n,
-                "TotalReads" => bucket.TotalReads,
-                "TotalWrites" => bucket.TotalWrites,
-                "Sessions" => bucket.SessionCount,
-                _ => bucket.TotalCpu,
-            };
-        }
-
-        QueryStoreSlicer.UpdateMetric(label);
-
-        if (QueryStoreGrid.SelectedItem != null)
-            QueryStoreGrid_SelectionChanged(QueryStoreGrid, null!);
-    }
-
-    // ── Procedure Stats Slicer ──
-
-    private string _procStatsSlicerMetric = "TotalCpu";
-    private List<TimeSliceBucket>? _procStatsSlicerData;
-
-    private async System.Threading.Tasks.Task LoadProcStatsSlicerAsync()
-    {
-        try
-        {
-            var hoursBack = GetHoursBack();
-            DateTime? fromDate = null, toDate = null;
-            if (IsCustomRange)
-            {
-                var fromLocal = GetDateTimeFromPickers(FromDatePicker!, FromHourCombo, FromMinuteCombo);
-                var toLocal = GetDateTimeFromPickers(ToDatePicker!, ToHourCombo, ToMinuteCombo);
-                if (fromLocal.HasValue && toLocal.HasValue)
-                {
-                    fromDate = ServerTimeHelper.DisplayTimeToServerTime(fromLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                    toDate = ServerTimeHelper.DisplayTimeToServerTime(toLocal.Value, ServerTimeHelper.CurrentDisplayMode);
-                }
-            }
-
-            var data = await _dataService.GetProcStatsSlicerDataAsync(_serverId, hoursBack, fromDate, toDate);
-            _procStatsSlicerData = data;
-            _procStatsSlicerMetric = "TotalCpu";
-            var (slicerStart, slicerEnd) = GetSlicerTimeRange(hoursBack, fromDate, toDate);
-            if (data.Count > 0)
-                ProcStatsSlicer.LoadData(data, "Total CPU (ms)", slicerStart, slicerEnd);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] LoadProcStatsSlicerAsync failed: {ex.Message}");
-        }
-    }
-
-    private async void OnProcStatsSlicerChanged(object? sender, Controls.SlicerRangeEventArgs e)
-    {
-        try
-        {
-            var fromServer = ServerTimeHelper.ToServerTime(e.StartUtc);
-            var toServer = ServerTimeHelper.ToServerTime(e.EndUtc);
-            var procStats = await _dataService.GetTopProceduresByCpuAsync(_serverId, 0, 50, fromServer, toServer, UtcOffsetMinutes);
-            _procStatsFilterMgr!.UpdateData(procStats);
-            await RefreshProcStatsComparisonAsync(fromServer, toServer);
-        }
-        catch (Exception ex)
-        {
-            AppLogger.Info("ServerTab", $"[{_server.DisplayName}] OnProcStatsSlicerChanged failed: {ex.Message}");
-        }
-    }
-
-    private void ProcedureStatsGrid_Sorting(object sender, DataGridSortingEventArgs e)
-    {
-        if (_procStatsSlicerData == null || _procStatsSlicerData.Count == 0) return;
-
-        var col = e.Column.SortMemberPath ?? "";
-        if (string.IsNullOrEmpty(col) && e.Column is DataGridBoundColumn bc && bc.Binding is System.Windows.Data.Binding b)
-            col = b.Path.Path;
-
-        var (metric, label) = col switch
-        {
-            "TotalCpuMs" => ("TotalCpu", "Total CPU (ms)"),
-            "AvgCpuMs" => ("AvgCpu", "Avg CPU (ms)"),
-            "TotalElapsedMs" => ("TotalElapsed", "Total Duration (ms)"),
-            "AvgElapsedMs" => ("AvgElapsed", "Avg Duration (ms)"),
-            "TotalLogicalReads" or "AvgReads" => ("TotalReads", "Total Reads"),
-            "TotalLogicalWrites" => ("TotalWrites", "Total Writes"),
-            "TotalPhysicalReads" => ("TotalReads", "Total Physical Reads"),
-            _ => ("TotalCpu", "Total CPU (ms)"),
-        };
-
-        if (metric == _procStatsSlicerMetric) return;
-        _procStatsSlicerMetric = metric;
-
-        foreach (var bucket in _procStatsSlicerData)
-        {
-            var n = bucket.SessionCount > 0 ? bucket.SessionCount : 1;
-            bucket.Value = metric switch
-            {
-                "TotalCpu" => bucket.TotalCpu,
-                "AvgCpu" => bucket.TotalCpu / n,
-                "TotalElapsed" => bucket.TotalElapsed,
-                "AvgElapsed" => bucket.TotalElapsed / n,
-                "TotalReads" => bucket.TotalReads,
-                "TotalWrites" => bucket.TotalWrites,
-                _ => bucket.TotalCpu,
-            };
-        }
-
-        ProcStatsSlicer.UpdateMetric(label);
-
-        if (ProcedureStatsGrid.SelectedItem != null)
-            ProcedureStatsGrid_SelectionChanged(ProcedureStatsGrid, null!);
     }
 
     private async void LiveSnapshot_Click(object sender, RoutedEventArgs e)
@@ -1573,22 +573,5 @@ public partial class ServerTab : UserControl
         _memoryPressureEventsHover?.Dispose();
         _currentWaitsDurationHover?.Dispose();
         _currentWaitsBlockedHover?.Dispose();
-    }
-
-    private static void SetDefaultSortIfNone(DataGrid grid, string bindingPath, ListSortDirection direction)
-    {
-        if (grid.Items.SortDescriptions.Count > 0) return;
-
-        grid.Items.SortDescriptions.Add(new SortDescription(bindingPath, direction));
-        foreach (var column in grid.Columns)
-        {
-            if (column is DataGridBoundColumn bc &&
-                bc.Binding is Binding b &&
-                b.Path.Path == bindingPath)
-            {
-                column.SortDirection = direction;
-                return;
-            }
-        }
     }
 }
